requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
langchain-core==1.2.20
langchain-community==0.4.1
langchain-openai==1.1.11
neo4j==5.28.1
chromadb==0.6.3
sentence-transformers==3.4.1
gitpython==3.1.44
# tree-sitter-languages (py-tree-sitter-languages) currently breaks with tree-sitter>=0.22
tree-sitter==0.21.3
tree-sitter-languages==1.10.2

